refactor(cli): abstract captureOutput to package
Use internal/capturer

ALLY-844
