+++ cli/cmd/vulnerability_test.go
@@ -27,10 +27,11 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/lacework/go-sdk/api"
+	"github.com/lacework/go-sdk/internal/capturer"
 )
 
 func TestBuildVulnContainerAssessmentReportsNoVulnerabilities(t *testing.T) {
-	cliOutput := captureOutput(func() {
+	cliOutput := capturer.CaptureOutput(func() {
 		assert.Nil(t, buildVulnContainerAssessmentReports(new(api.VulnContainerAssessment)))
 	})
 	assert.Contains(t, cliOutput, "Great news! This container image has no vulnerabilities...")
@@ -38,7 +39,7 @@ func TestBuildVulnContainerAssessmentReportsNoVulnerabilities(t *testing.T) {
 	t.Run("test JSON output", func(t *testing.T) {
 		cli.EnableJSONOutput()
 		defer cli.EnableHumanOutput()
-		cliJSONOutput := captureOutput(func() {
+		cliJSONOutput := capturer.CaptureOutput(func() {
 			assert.Nil(t, buildVulnContainerAssessmentReports(new(api.VulnContainerAssessment)))
 		})
 		expectedJSON := `{
@@ -56,7 +57,7 @@ func TestBuildVulnContainerAssessmentReportsNoVulnerabilities(t *testing.T) {
 }
 
 func TestBuildVulnContainerAssessmentReportsWithVulnerabilitiesButNoImageInformation(t *testing.T) {
-	cliOutput := captureOutput(func() {
+	cliOutput := capturer.CaptureOutput(func() {
 		err := buildVulnContainerAssessmentReports(&api.VulnContainerAssessment{TotalVulnerabilities: 10})
 		if assert.NotNil(t, err, "we would panic, please check this error") {
 			assert.Equal(t, "unable to build container vulnerability report. (image information not found)", err.Error())
@@ -69,7 +70,7 @@ func TestBuildVulnContainerAssessmentReportsWithVulnerabilitiesPackagesView(t *t
 	vulCmdState.Packages = true
 	defer clearVulnFilters()
 
-	cliOutput := captureOutput(func() {
+	cliOutput := capturer.CaptureOutput(func() {
 		assert.Nil(t, buildVulnContainerAssessmentReports(mockVulnerabilityAssessment()))
 	})
 	// NOTE (@afiune): We purposly leave trailing spaces in this table, we need them!
@@ -99,7 +100,7 @@ func TestBuildVulnContainerAssessmentReportsWithVulnerabilitiesPackagesViewWithF
 	vulCmdState.Packages = true
 	defer clearVulnFilters()
 
-	cliOutput := captureOutput(func() {
+	cliOutput := capturer.CaptureOutput(func() {
 		assert.Nil(t, buildVulnContainerAssessmentReports(mockVulnerabilityAssessment()))
 	})
 	// NOTE (@afiune): We purposly leave trailing spaces in this table, we need them!
@@ -124,7 +125,7 @@ func TestBuildVulnContainerAssessmentReportsWithVulnerabilitiesPackagesViewWithF
 	t.Run("test JSON output", func(t *testing.T) {
 		cli.EnableJSONOutput()
 		defer cli.EnableHumanOutput()
-		cliJSONOutput := captureOutput(func() {
+		cliJSONOutput := capturer.CaptureOutput(func() {
 			assert.Nil(t, buildVulnContainerAssessmentReports(mockVulnerabilityAssessment()))
 		})
 		expectedJSON := `{
