+++ lwlogger/logger_test.go
@@ -29,12 +29,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 
+	"github.com/lacework/go-sdk/internal/capturer"
 	"github.com/lacework/go-sdk/lwlogger"
 )
 
 func TestLoggerNew(t *testing.T) {
 	// by default we do not log an INFO or DEBUG, only ERROR
-	logOutput := captureOutput(func() {
+	logOutput := capturer.CaptureOutput(func() {
 		lwL := lwlogger.New("")
 		lwL.Info("interesting info")
 		lwL.Debug("we are debugging")
@@ -69,7 +70,7 @@ func TestLoggerNew(t *testing.T) {
 
 func TestLoggerNewINFO(t *testing.T) {
 	// switching to INFO will show ERRORS and INFO logs
-	logOutput := captureOutput(func() {
+	logOutput := capturer.CaptureOutput(func() {
 		lwL := lwlogger.New("INFO")
 		lwL.Info("interesting info")
 		lwL.Debug("we are debugging")
@@ -110,7 +111,7 @@ func TestLoggerNewINFO(t *testing.T) {
 
 func TestLoggerNewDEBUG(t *testing.T) {
 	// switching to DEBUG will show INFO, DEBUG and ERRORS messages
-	logOutput := captureOutput(func() {
+	logOutput := capturer.CaptureOutput(func() {
 		lwL := lwlogger.New("DEBUG")
 		lwL.Info("interesting info")
 		lwL.Debug("we are debugging")
@@ -160,7 +161,7 @@ func TestLoggerNewLogEnv(t *testing.T) {
 	os.Setenv(lwlogger.LogLevelEnv, "DEBUG")
 	defer os.Setenv(lwlogger.LogLevelEnv, "")
 
-	logOutput := captureOutput(func() {
+	logOutput := capturer.CaptureOutput(func() {
 		lwL := lwlogger.New("")
 		lwL.Info("interesting info")
 		lwL.Debug("we are debugging")
@@ -179,7 +180,7 @@ func TestLoggerNewLogFormatEnv(t *testing.T) {
 	os.Setenv(lwlogger.LogFormatEnv, "CONSOLE")
 	defer os.Setenv(lwlogger.LogFormatEnv, "")
 
-	logOutput := captureOutput(func() {
+	logOutput := capturer.CaptureOutput(func() {
 		lwL := lwlogger.New("DEBUG")
 		lwL.Info("interesting info")
 		lwL.Debug("we are debugging")
@@ -208,7 +209,7 @@ func TestLoggerNewWithWriter(t *testing.T) {
 }
 
 func TestLoggerNewWithOptions(t *testing.T) {
-	logOutput := captureOutput(func() {
+	logOutput := capturer.CaptureOutput(func() {
 		lwL := lwlogger.New("DEBUG",
 			zap.Fields(zap.Field(zap.String("my_field", "awesome"))),
 		)
@@ -227,9 +228,9 @@ func TestValidLevel(t *testing.T) {
 	assert.False(t, lwlogger.ValidLevel("FOO"))
 }
 
-// captureOutput executes a function and captures the STDOUT and STDERR,
+// capturer.CaptureOutput executes a function and captures the STDOUT and STDERR,
 // useful to test logging messages
-func captureOutput(f func()) string {
+func capturer.CaptureOutput(f func()) string {
 	r, w, err := os.Pipe()
 	if err != nil {
 		panic(err)
